batbot7_bringup/requirements.txt
# Author: Ben Westcott
# Created: 4/22/23

matplotlib
numpy
pyserial
pyyaml
pygnssutils
gpxpy
scipy
pynmeagps
pyrtcm
pyubx2
paho-mqtt
PyQt6
qdarkstyle
pyqtgraph